security-guardrails/requirements.txt
# Core dependencies
pyyaml==6.0.1
requests==2.31.0
github3.py==4.0.1
python-dotenv==1.0.0

# Security tools (will be installed)
# These are installed via system, not pip
# semgrep
# gitleaks
# trivy
# checkov

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Code quality
black==23.12.0
flake8==6.1.0
mypy==1.7.0